create load_and_parse_tfrecords for tests
